agents/google-search/pyproject.toml
[project]
name = "google-search"
version = "0.1.0"
description = "A google search agent"
authors = [
    {name = "Weiyuan Liu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
